// apps/client/src/app/core/swagger/model/statistics.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {StatisticsSample} from './statisticsSample';

export interface Statistics {
  sampling: string;
  samples: Array<StatisticsSample>;
  totalReports: number;
  avgReportsPerSample: number;
}