// apps/client/src/app/core/swagger/model/project.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {Environment} from './environment';

export interface Project {
  projectId: string;
  name: string;
  description: string;
  createDate: Date;
  environments: Array<Environment>;
}